pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-green-background"
version = "0.1.0"
description = "Automatic green-screen background tool with pywebview GUI."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "numpy",
  "opencv-python",
  "pywebview",
]

[project.scripts]
auto-green-background = "src.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
